Cleanup signal sending
Most notably, make preparation of plaintext for CMAC calculation
more readable.

+++ macan/include/cryptlib.h
@@ -32,8 +32,8 @@ void macan_aes_encrypt(const struct macan_key *key, size_t len, uint8_t *dst, co
 void macan_aes_decrypt(const struct macan_key *key, size_t len, uint8_t *dst, const uint8_t *src);
 void macan_aes_wrap(const struct macan_key *key, size_t length, uint8_t *dst, const uint8_t *src);
 int macan_aes_unwrap(const struct macan_key *key, size_t length, uint8_t *dst, uint8_t *src, uint8_t *tmp);
-int macan_check_cmac(struct macan_ctx *ctx, struct macan_key *skey, const uint8_t *cmac4, uint8_t *plain, uint8_t *fill_time, uint8_t len);
-void macan_sign(struct macan_key *skey, uint8_t *cmac4, uint8_t *plain, uint8_t len);
+int macan_check_cmac(struct macan_ctx *ctx, struct macan_key *skey, const uint8_t *cmac4, uint8_t *plain, uint8_t *fill_time, unsigned len);
+void macan_sign(struct macan_key *skey, uint8_t *cmac4, uint8_t *plain, unsigned len);
 void macan_unwrap_key(const struct macan_key *key, size_t srclen, uint8_t *dst, uint8_t *src);
 
 #endif /* CRYPTLIB_H */
